devtools/server/tests/chrome/test_inspector-reload.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gInspectee = null;
let gWalker = null;

addTest(async function setup() {
  const url = document.getElementById("inspectorContent").href;
  const { target, doc } = await attachURL(url);
  const inspector = await target.getFront("inspector");
  gInspectee = doc;
  const walker = inspector.walker;
  gWalker = await inspector.getWalker();

  ok(walker === gWalker, "getWalker() twice should return the same walker.");
  runNextTest();
});

addTest(async function testReload() {
  const oldRootID = gWalker.rootNode.actorID;

  // watchRootNodeResolve will be assigned to different promises throughout the
  // test to be able to await on individual calls of the watchRootNode callback.
  let watchRootNodeResolve;

  info("Start watching for root nodes and wait for the initial root node");
  const onInitialRootNodePromise = new Promise(r => (watchRootNodeResolve = r));
  const onInitialNewRootAvailable = gWalker.once("root-available");
  gWalker.watchRootNode(() => watchRootNodeResolve());
  await Promise.all([onInitialNewRootAvailable, onInitialRootNodePromise]);

  info("Retrieve the node front for the selector `#a`");
  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "#a");
  ok(nodeFront.actorID, "Node front has a valid actor ID");

  info("Reload the page and wait for the newRoot mutation");
  const onNewRootNodePromise = new Promise(r => (watchRootNodeResolve = r));
  const onNewRootAvailable = gWalker.once("root-available");

  gInspectee.defaultView.location.reload();
  await Promise.all([onNewRootAvailable, onNewRootNodePromise]);

  info("Retrieve the (new) node front for the selector `#a`");
  const newNodeFront = await gWalker.querySelector(gWalker.rootNode, "#a");
  ok(newNodeFront.actorID, "Got a new actor ID");
  ok(gWalker.rootNode.actorID != oldRootID, "Root node should have changed.");

  runNextTest();
});

addTest(function cleanup() {
  gWalker = null;
  gInspectee = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
